feat: update budget line items and enhance agreement handling in helpers
