+++ backend/data_tools/data/agreements_and_blin_data.json5
@@ -16535,11 +16535,10 @@
   },
     {
       "budget_line_item_type": "CONTRACT",
-      line_description: "BLI with no date needed #1",
+      line_description: "BLI with no date needed and agreement #1",
       comments: "",
       can_id: 520,
       created_by: 503,
-      agreement_id: 6,
       amount: 4162025.1,
       status: "DRAFT",
       on_hold: false,
@@ -16548,11 +16547,10 @@
     },
     {
       "budget_line_item_type": "CONTRACT",
-      line_description: "BLI with no date needed #2",
+      line_description: "BLI with no date needed and agreement #2",
       comments: "",
       can_id: 520,
       created_by: 503,
-      agreement_id: 6,
       amount: 4162025.2,
       status: "DRAFT",
       on_hold: false,
@@ -16561,7 +16559,7 @@
     },
     {
       "budget_line_item_type": "CONTRACT",
-      line_description: "BLI with no date needed #3",
+      line_description: "BLI with no date needed #1",
       comments: "",
       can_id: 520,
       created_by: 503,
@@ -16581,6 +16579,19 @@
       agreement_id: 6,
       amount: 4162025.0,
       status: "PLANNED",
+      date_needed: "2022-04-16T14:03:25.765487",
+      on_hold: false,
+      certified: false,
+      closed: false,
+    },
+     {
+      "budget_line_item_type": "CONTRACT",
+      line_description: "BLI with date needed without agreement",
+      comments: "",
+      can_id: 520,
+      created_by: 503,
+      amount: 4172025.0,
+      status: "DRAFT",
       date_needed: "2022-04-16T14:03:25.765487",
       on_hold: false,
       certified: false,
