+++ backend/ops_api/tests/ops/can/test_portfolio_cans.py
@@ -86,7 +86,7 @@ def test_bli_with_null_date_needed(app, auth_client):
 
     budget_line_items = [app.db_session.get(BudgetLineItem, bli_id) for bli_id in budget_line_item_ids]
 
-    assert len(budget_line_items) == 4
+    assert len(budget_line_items) == 5
     items_with_date = [bli for bli in budget_line_items if bli.date_needed is not None]
     items_without_date = [bli for bli in budget_line_items if bli.date_needed is None]
 
@@ -95,7 +95,7 @@ def test_bli_with_null_date_needed(app, auth_client):
     assert sum(bli.amount for bli in items_without_date) == Decimal("12486075.60")
     assert all(bli.status == BudgetLineItemStatus.DRAFT for bli in items_without_date)
 
-    assert len(items_with_date) == 1
+    assert len(items_with_date) == 2
     assert all(bli.date_needed is not None for bli in items_with_date)
-    assert sum(bli.amount for bli in items_with_date) == Decimal("4162025.0")
-    assert all(bli.status == BudgetLineItemStatus.PLANNED for bli in items_with_date)
+    assert sum(bli.amount for bli in items_with_date) == Decimal("4162025.0") + Decimal("4172025")
+    assert all(bli.status in [BudgetLineItemStatus.PLANNED, BudgetLineItemStatus.DRAFT] for bli in items_with_date)
