+++ backend/ops_api/tests/ops/funding_summary/test_can_funding_summary.py
@@ -227,7 +227,7 @@ def test_get_can_funding_summary_no_fiscal_year(loaded_db, test_can) -> None:
                         15867,
                         15880,
                         15921,
-                        16023,
+                        16024,
                     ],
                     "created_by": None,
                     "created_by_user": None,
@@ -370,7 +370,7 @@ def test_get_can_funding_summary_with_fiscal_year(loaded_db, test_can) -> None:
                         15867,
                         15880,
                         15921,
-                        16023,
+                        16024,
                     ],
                     "created_by": None,
                     "created_by_user": None,
