+++ frontend/src/helpers/budgetLines.helpers.js
@@ -1,3 +1,4 @@
+import { getTypesCounts } from "../pages/cans/detail/Can.helpers";
 import { formatDateToMonthDayYear } from "./utils";
 /**
  * @typedef {Object} BudgetLine
@@ -168,4 +169,28 @@ export const isBudgetLineEditableByStatus = (budgetLine) => {
     const isBudgetLineInReview = budgetLine?.in_review;
 
     return (isBudgetLineDraft || isBudgetLinePlanned) && !isBudgetLineInReview;
+};
+/**
+ * @typedef ItemCount
+ * @property {string} type
+ * @property {number} count
+ */
+/**
+ * @param {import("../components/BudgetLineItems/BudgetLineTypes").BudgetLine[]} budgetlines
+ * @returns {ItemCount[]}
+ */
+export const getAgreementTypesCount = (budgetlines) => {
+    const budgetLinesAgreements = budgetlines?.filter((item) => item.agreement).map((item) => item.agreement);
+    const uniqueBudgetLineAgreements =
+        budgetLinesAgreements?.reduce((acc, item) => {
+            // Skip if item is null or doesn't have a name
+            if (!item?.name) return acc;
+
+            if (!acc.some((existingItem) => existingItem?.name === item.name)) {
+                acc.push(item);
+            }
+            return acc;
+        }, []) ?? [];
+    const agreementTypesCount = getTypesCounts(uniqueBudgetLineAgreements ?? [], "agreement_type");
+    return agreementTypesCount;
 };
